Extract useCollection composable
